Add current progress fixing snake

+++ src/GameLibs/Snake/System/EatSystem.hpp
@@ -13,7 +13,7 @@ namespace Snake {
     namespace System {
         class EatSystem : public Arcade::ECS::ISystem {
             public:
-                void run(float deltaTime,
+                void run(double deltaTime,
                     Arcade::ECS::IEventManager &eventManager,
                     Arcade::ECS::IEntityManager &currentEntityManager) final;
         };
